# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AnalysisInput {
  /** The website URL to analyze */
  url: string;
  /**
     * Optional competitor URL
     * @nullable
     */
  competitorUrl?: string | null;
}

export type AnalysisSummaryStatus = typeof AnalysisSummaryStatus[keyof typeof AnalysisSummaryStatus];


export const AnalysisSummaryStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface AnalysisSummary {
  id: number;
  url: string;
  status: AnalysisSummaryStatus;
  /** Overall website score 0-100 */
  overallScore: number;
  createdAt: string;
}

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Analysis {
  id: number;
  url: string;
  /** @nullable */
  competitorUrl?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  seoScore?: number | null;
  /** @nullable */
  uxScore?: number | null;
  /** @nullable */
  contentScore?: number | null;
  /** @nullable */
  growthScore?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ScoreDashboard {
  overallScore: number;
  seoScore: number;
  uxScore: number;
  contentScore: number;
  growthScore: number;
  /** e.g. "+25% organic traffic in 90 days" */
  estimatedGrowthImpact: string;
}

export type KeywordIntent = typeof KeywordIntent[keyof typeof KeywordIntent];


export const KeywordIntent = {
  informational: 'informational',
  navigational: 'navigational',
  commercial: 'commercial',
  transactional: 'transactional',
} as const;

export type KeywordType = typeof KeywordType[keyof typeof KeywordType];


export const KeywordType = {
  primary: 'primary',
  'long-tail': 'long-tail',
  opportunity: 'opportunity',
} as const;

export interface Keyword {
  keyword: string;
  /** Relevance/opportunity score 0-100 */
  score: number;
  intent: KeywordIntent;
  type: KeywordType;
  /** @nullable */
  volume?: string | null;
}

export interface KeywordAnalysis {
  seoScore: number;
  keywords: Keyword[];
  opportunities: string[];
}

export interface Competitor {
  name: string;
  url: string;
  strengths: string[];
  weaknesses: string[];
  ctaScore: number;
  contentScore: number;
  uxScore: number;
}

export interface CompetitorAnalysis {
  competitors: Competitor[];
  yourAdvantages: string[];
  yourWeaknesses: string[];
}

export type ContentSuggestionPriority = typeof ContentSuggestionPriority[keyof typeof ContentSuggestionPriority];


export const ContentSuggestionPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ContentSuggestion {
  /** e.g. "Headline", "CTA", "Readability" */
  area: string;
  currentIssue: string;
  suggestedImprovement: string;
  expectedImpact: string;
  priority: ContentSuggestionPriority;
}

export interface ContentAnalysis {
  contentScore: number;
  readabilityScore: number;
  suggestions: ContentSuggestion[];
}

export interface FlowStep {
  step: number;
  label: string;
  score: number;
  issues: string[];
}

export interface UserFlowAnalysis {
  uxScore: number;
  mobileScore: number;
  accessibilityScore: number;
  flowSteps: FlowStep[];
  frictionPoints: string[];
  improvements: string[];
}

export type RecommendationCategory = typeof RecommendationCategory[keyof typeof RecommendationCategory];


export const RecommendationCategory = {
  seo: 'seo',
  content: 'content',
  ux: 'ux',
  growth: 'growth',
  technical: 'technical',
} as const;

export type RecommendationImpact = typeof RecommendationImpact[keyof typeof RecommendationImpact];


export const RecommendationImpact = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type RecommendationEffort = typeof RecommendationEffort[keyof typeof RecommendationEffort];


export const RecommendationEffort = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Recommendation {
  rank: number;
  title: string;
  description: string;
  category: RecommendationCategory;
  impact: RecommendationImpact;
  effort: RecommendationEffort;
}