# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from './analysisStatus';

export interface Analysis {
  id: number;
  url: string;
  /** @nullable */
  competitorUrl?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  seoScore?: number | null;
  /** @nullable */
  uxScore?: number | null;
  /** @nullable */
  contentScore?: number | null;
  /** @nullable */
  growthScore?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}