# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisInput {
  /** The website URL to analyze */
  url: string;
  /**
     * Optional competitor URL
     * @nullable
     */
  competitorUrl?: string | null;
}