# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSummaryStatus } from './analysisSummaryStatus';

export interface AnalysisSummary {
  id: number;
  url: string;
  status: AnalysisSummaryStatus;
  /** Overall website score 0-100 */
  overallScore: number;
  createdAt: string;
}