# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Competitor {
  name: string;
  url: string;
  strengths: string[];
  weaknesses: string[];
  ctaScore: number;
  contentScore: number;
  uxScore: number;
}