# lib/api-zod/src/generated/types/competitorAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { Competitor } from './competitor';

export interface CompetitorAnalysis {
  competitors: Competitor[];
  yourAdvantages: string[];
  yourWeaknesses: string[];
}