# lib/api-zod/src/generated/types/contentSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentSuggestionPriority } from './contentSuggestionPriority';

export interface ContentSuggestion {
  /** e.g. "Headline", "CTA", "Readability" */
  area: string;
  currentIssue: string;
  suggestedImprovement: string;
  expectedImpact: string;
  priority: ContentSuggestionPriority;
}