# lib/api-zod/src/generated/types/keyword.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { KeywordIntent } from './keywordIntent';
import type { KeywordType } from './keywordType';

export interface Keyword {
  keyword: string;
  /** Relevance/opportunity score 0-100 */
  score: number;
  intent: KeywordIntent;
  type: KeywordType;
  /** @nullable */
  volume?: string | null;
}