# lib/api-zod/src/generated/types/keywordAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { Keyword } from './keyword';

export interface KeywordAnalysis {
  seoScore: number;
  keywords: Keyword[];
  opportunities: string[];
}