# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationCategory } from './recommendationCategory';
import type { RecommendationEffort } from './recommendationEffort';
import type { RecommendationImpact } from './recommendationImpact';

export interface Recommendation {
  rank: number;
  title: string;
  description: string;
  category: RecommendationCategory;
  impact: RecommendationImpact;
  effort: RecommendationEffort;
}