# lib/api-zod/src/generated/types/scoreDashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreDashboard {
  overallScore: number;
  seoScore: number;
  uxScore: number;
  contentScore: number;
  growthScore: number;
  /** e.g. "+25% organic traffic in 90 days" */
  estimatedGrowthImpact: string;
}