# lib/api-zod/src/generated/types/userFlowAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Website Strategy & Growth Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { FlowStep } from './flowStep';

export interface UserFlowAnalysis {
  uxScore: number;
  mobileScore: number;
  accessibilityScore: number;
  flowSteps: FlowStep[];
  frictionPoints: string[];
  improvements: string[];
}